If a metric exists in tsidCache, register it in per-date index right away instead of deferring the registration until the updatePerDateData() is called.
Previously updatePerDateData() was used to add the metrics to the per-date index
in batch mode. This logic is not removed though since it is used when prefilling
the next day index.

the Storage.add() method has also been changed to not return an error. The only
case when it previously returned an error was the error returned by
updatePerDateData() and it was an error related to failure to unmarshal the
metricNameRaw. However, in all other places this error treated as warning.
So instead, the Storage.add() function now returns the number of rows that were
actually inserted. A number of metrics have been added to record the counts of
rows dropped due to unmarshalling errors and more. A unit test was added to
cover all cases then rows are dropped and not inserted.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>